cd: Introduce new binary build workflows (cpu)
Introduces new binary build workflows and some supplementary changes to
downstream scripts in order to accommodate the new workflows.

Goal here is to get off the ground with the new syntax and ideally make
it easier to add on more workflows after this one is solidified.

I'm not entirely happy with the state of some of these workflows since
there are lot of unused variables but I do want to test this as a POC
ASAP so we can flesh out the UX / how we'd like to maintain it.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>

ghstack-source-id: 1e83c3a5a9bdd3edccbc8e1dd020a8a11632cc00
Pull Request resolved: https://github.com/pytorch/pytorch/pull/150713

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>